requirements.txt
xgboost==1.4.2
matplotlib==3.4.2
numpy==1.19.5
streamlit==0.84.2
pandas==1.2.5
Pillow==8.3.1
scikit_learn==0.24.2
lasio==0.29
pickleshare==0.7.5